requirements.txt
pytest==8.4
selenium==4.36
allure-pytest==2.15